% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_MI.R
\name{comb_MI}
\alias{comb_MI}
\title{Combine Results from Multiple Imputation}
\usage{
comb_MI(coefmi, varmi, M)
}
\arguments{
\item{coefmi}{A numeric vector of coefficients from each imputation.}

\item{varmi}{A numeric vector of variances from each imputation.}

\item{M}{The number of imputations.}
}
\value{
A numeric vector containing:
\item{est}{The combined estimate (mean of the imputed coefficients).}
\item{se}{The combined standard error.}
\item{lowerCI}{The lower bound of the 95\% confidence interval.}
\item{upperCI}{The upper bound of the 95\% confidence interval.}
\item{pval}{The p-value for the hypothesis test of the combined estimate being zero.}
}
\description{
This function combines estimates and variances from multiple imputation using Rubin's rules. It calculates the combined estimate,
standard error, confidence intervals, and p-value based on the imputed coefficient estimates and variances.
}
\keyword{internal}
