% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipwR.R
\name{ipwR}
\alias{ipwR}
\title{Calculate Inverse Probability Weights (IPW) for Longitudinal Response Variable}
\usage{
ipwR(data1, ipwmodel, impmodel, idvar)
}
\arguments{
\item{data1}{A data frame containing the data for the longitudinal model.}

\item{ipwmodel}{A formula specifying the model for the inverse probability weights, typically a logistic regression model (e.g., `~ x1 + x2`).}

\item{impmodel}{A formula specifying the imputation model for the response variable (e.g., `y ~ x1 + x2 + (1|id)`).}

\item{idvar}{A string specifying the name of the variable identifying the subjects (ID variable).}
}
\value{
A list containing:
\item{ipw.y}{A vector of the inverse probability weights for each observation.}
\item{fittedvalue.id}{A data frame with the fitted values for each subject ID.}
}
\description{
This function calculates inverse probability weights (IPW) for a longitudinal response variable based on a given imputation model and
IPW model. It uses a logistic regression model to estimate the probability of observing a non-missing value for the response variable,
and calculates the weights accordingly.
}
\details{
The function first creates a binary variable indicating whether the response variable is missing or observed. It then fits a
logistic regression model for the probability of observing a non-missing response using the specified IPW model. The function calculates
the inverse probability weights as the inverse of the fitted values. Finally, it aggregates the weights at the subject level and returns
the calculated IPW and fitted values.
}
\keyword{internal}
