% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmLOCF.R
\name{jmLOCF}
\alias{jmLOCF}
\title{Two stage Joint Model with LOCF(Last observation carry forward method)}
\usage{
jmLOCF(ldata, sdata, id, visitTime, timeDep, coxModel = NULL, model = "Cox")
}
\arguments{
\item{ldata}{A data frame containing the longitudinal data. This should include individual-level data with repeated measures over time.}

\item{sdata}{A data frame containing the survival data, including the time-to-event and event status variables.}

\item{id}{A string specifying the column name in both `ldata` and `sdata` identifying the individual subject (e.g., patient ID).}

\item{visitTime}{A string specifying the column name in `ldata` for the time points at which observations were made.}

\item{timeDep}{A vector of column names in `ldata` specifying the time-dependent biomarkers.}

\item{coxModel}{A formula representing the Cox proportional hazards model. It is required to specify the structure of the model with the time-dependent biomarker.}

\item{model}{A string specifying the type of survival model to use. The default is `"Cox"`, but `"Aalen"` can be specified for an Aalen additive model.}
}
\value{
A list containing the following components:
\item{res}{A summary of the fitted model.}
\item{coefs}{A data frame containing the coefficients from the fitted model.}
\item{vars}{A data frame containing the variance-covariance matrix of the coefficients.}
\item{nam}{The names of the coefficients.}
\item{fitted_model}{The fitted model object (either `coxph` or `aalen`).}
\item{imp.data}{A data frame containing the imputed longitudinal data.}
\item{complete_data}{A data frame containing the complete data with time-dependent biomarker imputations.}
\item{tname}{The name of the time variable used in the model.}
\item{ename}{The name of the event variable used in the model.}
\item{method}{The method used for imputations ("LOCF").}
}
\description{
This function fits a joint model using the Last Observation Carried Forward (LOCF) method for time-dependent biomarkers with missing values. It handles longitudinal data and survival data jointly using a Cox proportional hazards model (or an Aalen additive model). The function imputes missing values for time-dependent biomarkers using the LOCF method.
}
\details{
This function implements a joint modeling framework for longitudinal and survival data using the LOCF method for imputing missing time-dependent biomarkers. It supports both the Cox proportional hazards model (`model="Cox"`) and the Aalen additive model (`model="Aalen"`).

The longitudinal data is first pre-processed to create time-dependent covariates, and then the missing biomarker values are imputed using LOCF. The survival data is used to fit a joint model, where the survival model accounts for the time-dependent covariates.
}
\examples{
 library(survival)
 model_jmLOCF<-jmLOCF(ldata=long_data,sdata=surv_data,
       timeDep=c("marker_1","marker_2","marker_3"),
       visitTime="Time",
       coxModel=Surv(survival_time,survival_status)~Z_1+Z_2+td(marker_1)+
                     td(marker_2)+td(marker_3),
       model="Cox",
       id="ID")
 model_jmLOCF
}
\references{
Goodrich, B., et al. "rstanarm: Bayesian applied regression modeling via Stan. R package version 2.17. 4." Online< http://mc-stan. org (2018).

Bhattacharjee, A., Rajbongshi, B. K., & Vishwakarma, G. K. (2024). jmBIG: enhancing dynamic risk prediction and personalized medicine through joint modeling of longitudinal and survival data in big routinely collected data. BMC Medical Research Methodology, 24(1), 172.
}
\author{
Atanu Bhattacharjee, Bhrigu Kumar Rajbongshi and Gajendra Kumar Vishwakarma
}
