% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmwMI.R
\name{jmwMI}
\alias{jmwMI}
\title{Two stage Joint Model with MI(Multiple imputation) and IPW(Inverse probabilty weight)}
\usage{
jmwMI(
  ldata,
  sdata,
  M = 5,
  id,
  visitTime,
  timeDep,
  impModel = NULL,
  ipwModel = NULL,
  coxModel = NULL,
  model = "Cox"
)
}
\arguments{
\item{ldata}{A data frame containing the longitudinal data (biomarker measurements over time) with columns
including the subject ID, visit times, time-dependent covariates, and other relevant data.}

\item{sdata}{A data frame containing the survival data, including the time-to-event information and the event status.}

\item{M}{An integer specifying the number of imputations to generate. Default is 5.}

\item{id}{The name of the variable representing the unique individual ID in both `ldata` and `sdata`.}

\item{visitTime}{The name of the variable representing the visit time in the longitudinal data (`ldata`).}

\item{timeDep}{A character vector of the time-dependent variables (biomarkers) in `ldata` that will be imputed.}

\item{impModel}{A list of imputation models, one for each time-dependent variable in `timeDep`. Each model should
be specified using a formula.}

\item{ipwModel}{A list of models used for calculating inverse probability weights (IPWs) for the time-dependent
variables in `timeDep`. Each model should be specified using a formula.}

\item{coxModel}{A survival model formula for Cox proportional hazards regression, such as `Surv(time, event) ~ x1 + x2 + td(x3)`.}

\item{model}{A string specifying the type of survival model to fit. Options are `'Cox'` (default) or `'Aalen'`.}
}
\value{
A list containing:
  \describe{
    \item{res}{A summary of the results from the survival model.}
    \item{coefs}{A data frame containing the model coefficients.}
    \item{vars}{A data frame containing the variances of the model coefficients.}
    \item{nam}{A character vector containing the names of the model coefficients.}
    \item{fitted_model}{A list containing the fitted models for each imputation.}
    \item{imp.data}{A list of imputed data for the time-dependent variables.}
    \item{miss.data}{The missing data for the time-dependent variables in the original dataset.}
    \item{complete_data}{A list of completed datasets after imputation.}
    \item{tname}{The name of the time variable in the survival data.}
    \item{ename}{The name of the event variable in the survival data.}
    \item{method}{The method used for the analysis. Always "MIm".}
  }
}
\description{
The `jmwMI` function performs multiple imputation on a dataset containing
time-dependent covariates (TDCs) for survival analysis. It imputes missing
values in longitudinal biomarkers, applies inverse probability weighting (IPW)
for longitudinal variables, and fits a survival model (Cox or Aalen) using the
imputed datasets. The function returns the results of the analysis, including
model coefficients, variances, and imputed data.
}
\examples{
library(survival)
model_wMI<- jmwMI(ldata=long_data,sdata=surv_data,
                 timeDep=c("marker_1","marker_2","marker_3"),
                 impModel=list(marker_1~Z_1+Z_2+Time+(1|ID),
                               marker_2~Z_1+Z_2+Time+(1|ID),
                               marker_3~Z_1+Z_2+Time+(1|ID)),
                 ipwModel=list(~Z_1+Time+(1|ID),
                               ~Z_1+Time+(1|ID),
                               ~Z_1+Time+(1|ID)),
                 visitTime="Time",
                 coxModel=Surv(survival_time,survival_status)~Z_1+Z_2+
                 td(marker_1)+td(marker_2)+td(marker_3),
                 model="Cox",id="ID")
model_wMI
}
\references{
Goodrich, B., et al. "rstanarm: Bayesian applied regression modeling via Stan. R package version 2.17. 4." Online< http://mc-stan. org (2018).

Bhattacharjee, A., Rajbongshi, B. K., & Vishwakarma, G. K. (2024). jmBIG: enhancing dynamic risk prediction and personalized medicine through joint modeling of longitudinal and survival data in big routinely collected data. BMC Medical Research Methodology, 24(1), 172.
}
\author{
Atanu Bhattacharjee, Bhrigu Kumar Rajbongshi and Gajendra Kumar Vishwakarma
}
