% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCF.R
\name{LOCF}
\alias{LOCF}
\title{Last Observation Carried Forward (LOCF) Imputation}
\usage{
LOCF(dataset, marker)
}
\arguments{
\item{dataset}{A data frame or tibble that contains the data. The dataset should have at least one column with missing values to apply LOCF imputation.}

\item{marker}{The name (or index) of the column in the dataset where the missing values should be replaced with the last observed value.}
}
\value{
A data frame with the missing values in the `marker` column replaced by the last observed value before them.
}
\description{
This function performs Last Observation Carried Forward (LOCF) imputation for missing values in a specified column (marker) of a dataset.
The missing values are replaced by the last observed value before them in the dataset.
}
\examples{
data1 <- data.frame(id = 1:5, value = c(1, NA, 3, NA, 5))
LOCF(data1, "value")
}
