% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longmodeldata.R
\name{longmodeldata}
\alias{longmodeldata}
\title{Prepare Data for Longitudinal Model with Random Effects}
\usage{
longmodeldata(data, longformula, idvar = NULL)
}
\arguments{
\item{data}{A data frame containing the variables for the model.}

\item{longformula}{A formula specifying the longitudinal model with both fixed and random effects. The formula should include
the fixed effects on the right-hand side of the `~` and random effects enclosed in parentheses, for example:
`albumin ~ trt + age + day + (1|id) + (day|id)`.}

\item{idvar}{(Optional) A string specifying the name of the ID variable in the dataset. If not provided, the function assumes
that the ID variable is present in the formula.}
}
\value{
A list containing the following elements:
\item{id}{A vector with the ID variable from the dataset.}
\item{long.y}{A vector of the dependent variable (Y) for the longitudinal model.}
\item{long.x}{A matrix of the fixed effects (X) from the model formula.}
\item{long.z}{A list of model matrices for the random effect terms (Z).}
\item{varzfirst}{A list of the first variable names in the random effect terms.}
\item{varzsecond}{A list of the second variable names in the random effect terms.}
}
\description{
This function prepares data for a longitudinal model by extracting the necessary components from the formula and splitting the data
according to the random effect terms. The function handles both fixed and random effects, including random slopes and intercepts,
and returns the processed data in a list suitable for model fitting.
}
\keyword{internal}
