% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.jm2s}
\alias{print.jm2s}
\title{Print method for jm2s objects}
\usage{
\method{print}{jm2s}(x, ...)
}
\arguments{
\item{x}{An object of class `jm2s`, returned by the `jm2s` function.}

\item{...}{Additional arguments (not used).}
}
\value{
Printed summary of the `jm2s` object.
}
\description{
Print method for jm2s objects
}
