% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.jmCC}
\alias{print.jmCC}
\title{Print method for jmCC objects}
\usage{
\method{print}{jmCC}(x, ...)
}
\arguments{
\item{x}{An object of class `jmCC`, returned by the `jmCC` function.}

\item{...}{Additional arguments (not used).}
}
\value{
Printed summary of the `jmCC` object.
}
\description{
Print method for jmCC objects
}
