% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.jmLOCF}
\alias{print.jmLOCF}
\title{Print method for jmLOCF objects}
\usage{
\method{print}{jmLOCF}(x, ...)
}
\arguments{
\item{x}{An object of class `jmLOCF`, returned by the `jmLOCF` function.}

\item{...}{Additional arguments (not used).}
}
\value{
Printed summary of the `jmLOCF` object.
}
\description{
Print method for jmLOCF objects
}
