% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.jmMI}
\alias{print.jmMI}
\title{Print method for jmMI objects}
\usage{
\method{print}{jmMI}(x, ...)
}
\arguments{
\item{x}{An object of class `jmMI`, returned by the `jmMI` function.}

\item{...}{Additional arguments (not used).}
}
\value{
Printed summary of the `jmMI` object.
}
\description{
Print method for jmMI objects
}
