% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.jmwMI}
\alias{print.jmwMI}
\title{Print method for jmMIIPW objects}
\usage{
\method{print}{jmwMI}(x, ...)
}
\arguments{
\item{x}{An object of class `jmwMI`, returned by the `jmwMI` function.}

\item{...}{Additional arguments (not used).}
}
\value{
Printed summary of the `jmwMI` object.
}
\description{
Print method for jmMIIPW objects
}
