% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_aalen.R
\name{summ_aalen}
\alias{summ_aalen}
\title{Summarize Results from Aalen's Additive Regression Model}
\usage{
summ_aalen(fit)
}
\arguments{
\item{fit}{A fitted model object from Aalen's additive regression, typically obtained from the `aalen()` function.}
}
\value{
A data frame containing the following columns:
\item{HD}{The estimated hazard coefficient for each variable.}
\item{SE}{The standard error of the hazard coefficient.}
\item{CIlow}{The lower bound of the 95\% confidence interval for the hazard coefficient.}
\item{CIupp}{The upper bound of the 95\% confidence interval for the hazard coefficient.}
\item{p-value}{The p-value for testing the null hypothesis that the hazard coefficient is zero.}
}
\description{
This function extracts and summarizes the results from an Aalen's additive regression model fit. It calculates the estimates,
standard errors, 95% confidence intervals, and p-values for each coefficient, based on the model fit object.
}
\keyword{internal}
