% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_cox.R
\name{summ_cox}
\alias{summ_cox}
\title{Summarize Cox Proportional Hazards Model}
\usage{
summ_cox(fit)
}
\arguments{
\item{fit}{A fitted Cox proportional hazards model object (of class `coxph`), typically returned by the `coxph` function.}
}
\value{
A data frame with the following columns:
\item{logHR}{The log hazard ratio for each coefficient.}
\item{SE}{The standard error of each coefficient.}
\item{CIlow}{The lower bound of the 95\% confidence interval for each coefficient.}
\item{CIupp}{The upper bound of the 95\% confidence interval for each coefficient.}
\item{p-value}{The p-value for testing the null hypothesis that each coefficient is zero.}
}
\description{
This function takes the output of a Cox proportional hazards model (`coxph`) and returns a summary data frame containing
the log hazard ratio (logHR), standard error (SE), confidence intervals (CI), and p-values for the model coefficients.
}
\keyword{internal}
