% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethods.R
\name{summary.jm2s}
\alias{summary.jm2s}
\title{Summary method for jm2s objects}
\usage{
\method{summary}{jm2s}(object, ...)
}
\arguments{
\item{object}{An object of class `jm2s`, returned by the `jm2s` function.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary of the `jm2s` object, including method, coefficients, variance, and fitted model.
}
\description{
Summary method for jm2s objects
}
