% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethods.R
\name{summary.jmCC}
\alias{summary.jmCC}
\title{Summary method for jmCC objects}
\usage{
\method{summary}{jmCC}(object, ...)
}
\arguments{
\item{object}{An object of class `jmCC`, returned by the `jmCC` function.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary of the `jmCC` object.
}
\description{
Summary method for jmCC objects
}
