% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethods.R
\name{summary.jmLOCF}
\alias{summary.jmLOCF}
\title{Summary method for jmLOCF objects}
\usage{
\method{summary}{jmLOCF}(object, ...)
}
\arguments{
\item{object}{An object of class `jmLOCF`, returned by the `jmLOCF` function.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary of the `jmLOCF` object.
}
\description{
Summary method for jmLOCF objects
}
