% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethods.R
\name{summary.jmMI}
\alias{summary.jmMI}
\title{Summary method for jmMI objects}
\usage{
\method{summary}{jmMI}(object, ...)
}
\arguments{
\item{object}{An object of class `jmMI`, returned by the `jmMI` function.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary of the `jmMI` object.
}
\description{
Summary method for jmMI objects
}
