% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryMethods.R
\name{summary.jmwMI}
\alias{summary.jmwMI}
\title{Summary method for jmwMI objects}
\usage{
\method{summary}{jmwMI}(object, ...)
}
\arguments{
\item{object}{An object of class `jmwMIIPW`, returned by the `jmwMI` function.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary of the `jmwMI` object.
}
\description{
Summary method for jmwMI objects
}
