% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survmodeldata.R
\name{survmodeldata}
\alias{survmodeldata}
\title{Prepare Data for Survival Model}
\usage{
survmodeldata(data, survformula, idvar = NULL, timeVar = NULL)
}
\arguments{
\item{data}{A data frame containing the data for the survival model.}

\item{survformula}{A formula specifying the survival model, typically of the form `Surv(time, status) ~ predictors`.}

\item{idvar}{A string specifying the name of the variable identifying the subjects (ID variable). Default is `NULL`.}

\item{timeVar}{A string specifying the time variable to be used in the survival analysis. Default is `NULL`.}
}
\value{
A list containing:
\item{status}{A vector representing the event status variable from the survival data.}
\item{time}{A vector representing the time variable from the survival data.}
\item{time.var.long}{A vector representing the time variable for longitudinal data, if applicable.}
\item{x}{A matrix of covariates for the survival model (design matrix).}
\item{idvar}{The ID variable used to identify the subjects in the data.}
\item{idlist}{A vector of subject IDs from the data.}
}
\description{
This function prepares the data for fitting a survival model by extracting the relevant variables from a formula and returning a list
containing the variables for the survival analysis. It processes the survival formula, extracts the status and time variables, and
constructs the necessary data structure for further modeling.
}
\details{
The function takes a survival formula (e.g., `Surv(time, status) ~ age + trt`) and extracts the status and time variables from
the `Surv()` function. It then extracts the predictor variables from the formula and prepares the necessary data for the survival model.
}
\keyword{internal}
