% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survProb.R
\name{survProb}
\alias{survProb}
\title{Function for caluculating Baseline cummulative hazard}
\usage{
survProb(fitted.object, nimp = 1, conf_level = 0.95)
}
\arguments{
\item{fitted.object}{model object fitted through JMtwostage}

\item{nimp}{imputation number}

\item{conf_level}{confindence level}
}
\value{
list with baseline hazard, baseline cummulative hazard.
}
\description{
This function calculates the baseline hazard, baseline cummulative hazard, at different unique event times
}
\examples{
library(survival)
model_wMI<- jmwMI(ldata=long_data,sdata=surv_data,
                  timeDep=c("marker_1","marker_2","marker_3"),
                  impModel=list(marker_1~Z_1+Z_2+Time+(1|ID),
                                marker_2~Z_1+Z_2+Time+(1|ID),
                                marker_3~Z_1+Z_2+Time+(1|ID)),
                  ipwModel=list(~Z_1+Time+(1|ID),
                                ~Z_1+Time+(1|ID),
                                ~Z_1+Time+(1|ID)),
                  visitTime="Time",
                  coxModel=Surv(survival_time,survival_status)~Z_1+Z_2+
                           td(marker_1)+td(marker_2)+td(marker_3),
                  model="Cox",id="ID")
survProb(model_wMI,nimp=1)
}
