% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_2s.R
\name{td}
\alias{td}
\title{td: Identity Function (Internal)}
\usage{
td(x)
}
\arguments{
\item{x}{A vector or object that is returned unchanged.}
}
\value{
The same value as the input.
}
\description{
A simple identity function that returns the input as is. This function is typically used as a placeholder or for specific purposes like indicating time-dependent variables in model formulas.
}
\details{
The `td` function does not modify the input value and simply returns it. It can be used in situations where a time-dependent covariate is needed in model formulas, typically in survival analysis models.
}
\keyword{internal}
